did remaining renames (regarding analysis -> report) plus moved GameStatusAnalysis to test package
