+++ src/main/resources/com/dlb/chess/internationalization/messages.properties
@@ -248,8 +248,6 @@ validation.san.checkSymbolButNoCheck=The move is neither check nor checkmate, bu
 validation.san.noSymbolButCheckmate=The move is checkmate, but no symbol was specified.
 validation.san.noSymbolButCheck=The move is check, but no symbol was specified.
 
-##analzyer
-
 #repetitions
 report.repetition.threefold.ahead.title=Valid threefold claims ahead
 report.repetition.threefold.ahead.none=None
