+++ src/test/java/com/dlb/chess/test/winnable/model/GameForcedAnalysis.java
@@ -1,6 +1,6 @@
 package com.dlb.chess.test.winnable.model;
 
-import com.dlb.chess.common.enums.GameStatusAnalysis;
+import com.dlb.chess.test.winnable.enums.GameStatusAnalysis;
 
 public record GameForcedAnalysis(GameStatusAnalysis gameStatus, int numberOfForcedHalfMoves) {
 }
