+++ src/test/java/com/dlb/chess/test/winnable/model/GameMultipleAnalysis.java
@@ -3,7 +3,7 @@ package com.dlb.chess.test.winnable.model;
 import java.util.Set;
 
 import com.dlb.chess.board.enums.Side;
-import com.dlb.chess.common.enums.GameStatusAnalysis;
+import com.dlb.chess.test.winnable.enums.GameStatusAnalysis;
 
 public record GameMultipleAnalysis(Set<GameStatusAnalysis> gameStatusSet, int numberOfHalfMoves, Side havingMove) {
