+++ src/test/java/com/dlb/chess/test/winnable/model/WinnableAnalysis.java
@@ -2,7 +2,7 @@ package com.dlb.chess.test.winnable.model;
 
 import java.util.Set;
 
-import com.dlb.chess.common.enums.GameStatusAnalysis;
+import com.dlb.chess.test.winnable.enums.GameStatusAnalysis;
 import com.dlb.chess.test.winnable.enums.Winnable;
 
 public record WinnableAnalysis(Winnable winnable, Set<GameStatusAnalysis> gameStatusSet) {
